Added seed.js and updated server.js
